# dm/src/main/resources/oas/dm.yml
openapi: "3.0.0"
info:
  title: Data Management Application
  version: 1.0.0
servers:
  - url: http://localhost:8081/dm
paths:
  /client:
    post:
      summary: Запрос клиентов
      operationId: getClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientSearchDto"
        required: true
      responses:
        '200':
          description: Информации о клиенте найдена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientDto"
  /client/contact:
    post:
      summary: Запрос контактов клиента
      operationId: getContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientIdDto"
        required: true
      responses:
        '200':
          description: Информации о контактах найдена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactDto"
  /client/account:
    post:
      summary: Запрос счетов клиента
      operationId: getAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientIdDto"
        required: true
      responses:
        '200':
          description: Информации о счетах найдена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountDto"
  /client/account/balance:
    post:
      summary: Запрос баланса по счету
      operationId: getAccountBalance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountNumberDto"
        required: true
      responses:
        '200':
          description: Баланс по счету определен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentBalanceDto"
  /client/info:
    post:
      summary: Запрос клиента по Id
      operationId: getClientInfo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientIdDto"
        required: true
      responses:
        '200':
          description: Клиент по Id найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDto"
        '400':
          description: Клиент по Id найден
  /client/account/operation:
    post:
      summary: Запрос операций по счету
      operationId: getAccountOperations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationSearchDto"
        required: true
      responses:
        '200':
          description: Операции по счету найдены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationDto"
  /client/contact/save:
    post:
      summary: Сохранение контакта клиента
      operationId: saveContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactDto"
        required: true
      responses:
        '200':
          description: Контакт клиента сохранен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDto"
        '400':
          description: Клиент не найден
  /client/level:
    post:
      summary: Получение уровня клиента
      operationId: getClientLevel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientIdDto"
        required: true
      responses:
        '200':
          description: Уровень клиента определен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLevelDto"
        '400':
          description: Клиент не найден
  /client/account/loan-payment:
    post:
      summary: Получение суммы процентных платежей по счету Овердрафт
      operationId: getLoanPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountNumberDto"
        required: true
      responses:
        '200':
          description: Сумма процентных платежей определена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanPaymentDto"
  /access:
    post:
      summary: Запрос прав доступа для Роли
      operationId: getAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestDto"
        required: true
      responses:
        '200':
          description: Права досиупа определены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessDto"
components:
  schemas:
    ClientSearchDto:
      $ref: "./common-dto.yml#/components/schemas/ClientSearchDto"
    ClientIdDto:
      $ref: "./common-dto.yml#/components/schemas/ClientIdDto"
    ClientDto:
      $ref: "./common-dto.yml#/components/schemas/ClientDto"
    ContactDto:
      $ref: "./common-dto.yml#/components/schemas/ContactDto"
    AccountNumberDto:
      $ref: "./common-dto.yml#/components/schemas/AccountNumberDto"
    CurrentBalanceDto:
      $ref: "./common-dto.yml#/components/schemas/CurrentBalanceDto"
    AccountDto:
      $ref: "./common-dto.yml#/components/schemas/AccountDto"
    OperationSearchDto:
      $ref: "./common-dto.yml#/components/schemas/OperationSearchDto"
    OperationDto:
      $ref: "./common-dto.yml#/components/schemas/OperationDto"
    ClientLevelDto:
      $ref: "./common-dto.yml#/components/schemas/ClientLevelDto"
    LoanPaymentDto:
      $ref: "./common-dto.yml#/components/schemas/LoanPaymentDto"
    AccessRequestDto:
      $ref: "./dm-dto.yml#/components/schemas/AccessRequestDto"
    AccessDto:
      $ref: "./dm-dto.yml#/components/schemas/AccessDto"